fix(as-5344): move appeal-details-validation out of insert directory
