+++ packages/business-rules/src/schemas/components/appeal-details-validation/appeal-type/appeal-type-validation.js
@@ -1,0 +1,15 @@
+const pinsYup = require('../../../../lib/pins-yup');
+const stringValidation = require('../../string-validators/string-validation');
+const stringSelectionValidation = require('../../string-validators/string-selection-validation');
+const { APPEAL_ID } = require('../../../../constants');
+
+const appealTypeValidation = () => {
+	return pinsYup.lazy((appealType) => {
+		if (appealType) {
+			return stringSelectionValidation(Object.values(APPEAL_ID));
+		}
+		return stringValidation();
+	});
+};
+
+module.exports = appealTypeValidation;
