+++ packages/business-rules/src/schemas/components/appeal-details-validation/id/id-validation.js
@@ -1,4 +1,4 @@
-const pinsYup = require('../../../../../lib/pins-yup');
+const pinsYup = require('../../../../lib/pins-yup');
 
 const idValidation = () => {
 	return pinsYup.string().trim().uuid().required();
