+++ packages/business-rules/src/schemas/components/appeal-details-validation/planning-application-number/planning-application-number-validation.js
@@ -1,4 +1,4 @@
-const pinsYup = require('../../../../../lib/pins-yup');
+const pinsYup = require('../../../../lib/pins-yup');
 
 const planningApplicationNumberValidation = () => {
 	return pinsYup.string().max(30).nullable();
