+++ packages/business-rules/src/schemas/components/appeal-details-validation/state/state-validation.js
@@ -1,5 +1,5 @@
-const pinsYup = require('../../../../../lib/pins-yup');
-const { APPEAL_STATE } = require('../../../../../constants');
+const pinsYup = require('../../../../lib/pins-yup');
+const { APPEAL_STATE } = require('../../../../constants');
 
 const stateValidation = () => {
 	return pinsYup.string().oneOf(Object.values(APPEAL_STATE)).default(APPEAL_STATE.DRAFT);
